+++ git-repository/src/easy/mod.rs
@@ -29,7 +29,7 @@ pub mod commit;
 pub mod head;
 pub mod iter;
 pub mod object;
-mod oid;
+pub mod oid;
 pub mod reference;
 pub mod state;