+++ git-repository/src/easy/object/mod.rs
@@ -131,16 +131,16 @@ where
     /// As [`to_commit_iter()`][ObjectRef::to_commit_iter()] but panics if this is not a commit
     pub fn commit_iter(&self) -> CommitRefIter<'_> {
         git_odb::data::Object::new(self.kind, &self.data)
-            .into_commit_iter()
+            .try_into_commit_iter()
             .expect("BUG: This object must be a commit")
     }
 
     pub fn to_commit_iter(&self) -> Option<CommitRefIter<'_>> {
-        git_odb::data::Object::new(self.kind, &self.data).into_commit_iter()
+        git_odb::data::Object::new(self.kind, &self.data).try_into_commit_iter()
     }
 
     pub fn to_tag_iter(&self) -> Option<TagRefIter<'_>> {
-        git_odb::data::Object::new(self.kind, &self.data).into_tag_iter()
+        git_odb::data::Object::new(self.kind, &self.data).try_into_tag_iter()
     }
 }